remove duplicate runs empty state
