+++ frontend/src/pages/pipelines/global/runs/ActiveRuns.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
 import {
   Bullseye,
@@ -8,20 +8,17 @@ import {
   EmptyStateIcon,
   Spinner,
   EmptyStateHeader,
-  EmptyStateFooter,
-  EmptyStateActions,
-  Button,
 } from '@patternfly/react-core';
-import { CubesIcon, ExclamationCircleIcon, PlusCircleIcon } from '@patternfly/react-icons';
+import { CubesIcon, ExclamationCircleIcon } from '@patternfly/react-icons';
 
 import PipelineRunTable from '~/concepts/pipelines/content/tables/pipelineRun/PipelineRunTable';
 import { usePipelineActiveRunsTable } from '~/concepts/pipelines/content/tables/pipelineRun/usePipelineRunTable';
 import { createRunRoute } from '~/routes';
 import { useContextExperimentArchived } from '~/pages/pipelines/global/experiments/ExperimentContext';
+import { EmptyRunsState } from '~/concepts/pipelines/content/tables/pipelineRun/EmptyRunsState';
 import { PipelineRunTabTitle, PipelineRunType } from './types';
 
 export const ActiveRuns: React.FC = () => {
-  const navigate = useNavigate();
   const { namespace, experimentId, pipelineVersionId, pipelineId } = useParams();
   const [[{ items: runs, totalSize }, loaded, error], { initialLoaded, ...tableProps }] =
     usePipelineActiveRunsTable({ experimentId, pipelineVersionId });
@@ -70,33 +67,10 @@ export const ActiveRuns: React.FC = () => {
 
   if (loaded && totalSize === 0 && !tableProps.filter) {
     return (
-      <EmptyState data-testid="active-runs-empty-state">
-        <EmptyStateHeader
-          titleText="No runs"
-          icon={<EmptyStateIcon icon={PlusCircleIcon} />}
-          headingLevel="h2"
-        />
-
-        <EmptyStateBody>
-          To get started, create a run. Alternatively, go to the{' '}
-          <b>{PipelineRunTabTitle.SCHEDULES}</b> tab and create a schedule to execute recurring
-          runs.
-        </EmptyStateBody>
-
-        <EmptyStateFooter>
-          <EmptyStateActions>
-            <Button
-              data-testid="create-run-button"
-              variant="primary"
-              onClick={() =>
-                navigate(createRunRoute(namespace, experimentId, pipelineId, pipelineVersionId))
-              }
-            >
-              Create run
-            </Button>
-          </EmptyStateActions>
-        </EmptyStateFooter>
-      </EmptyState>
+      <EmptyRunsState
+        createRunRoute={createRunRoute(namespace, experimentId, pipelineId, pipelineVersionId)}
+        dataTestId="active-runs-empty-state"
+      />
     );
   }
